WFP-2684 moving database service into own file
